fix(transport): use pre-generated protobuf code for Windows builds (#4)
The protobuf-src crate fails to build on Windows due to abseil-cpp
linker issues with UCRT math functions (ceilf, ldexp, nan, etc.).

This commit pre-generates the protobuf code and checks it into the
repository, which:
- Eliminates the need for protoc at build time
- Enables builds on all platforms including Windows
- Improves build times by skipping protobuf compilation
- Follows the pattern used by tonic-health and other crates

The `regenerate-proto` feature is available for developers who need
to update the proto definitions.

Also fixes pre-existing Windows compilation issues:
- Type mismatch in max_instances (u32 -> usize)
- Dead code warning for NamedPipeState read_buffer

Closes #3

+++ crates/mcpkit-transport/src/windows.rs
@@ -132,6 +132,7 @@ impl NamedPipeConfig {
 }
 
 /// Internal state for reading/writing.
+#[allow(dead_code)]
 struct NamedPipeState {
     /// Line buffer for reading complete messages.
     line_buffer: String,
@@ -503,7 +504,7 @@ impl TransportListener for NamedPipeServer {
         // Create a new pipe instance for this connection
         let pipe = ServerOptions::new()
             .first_pipe_instance(false)
-            .max_instances(self.config.max_instances)
+            .max_instances(self.config.max_instances as usize)
             .in_buffer_size(self.config.in_buffer_size as u32)
             .out_buffer_size(self.config.out_buffer_size as u32)
             .create(&self.config.name)
